Refactor circuits forms
